Migrate to scene delegate
